Create themes.syd
repositiores will contain key value pairs where the key is the id of the theme. The value for the id is another key value store where url is the url of the repository and 'hosting' will help identify how to retrieve from the url (for example, by putting 'github' there we will know how to interact with it. Other example values for 'hosting': web (http or https), ftp, etc)